Add test for operations
