+++ tests/core/test_array.py
@@ -43,39 +43,6 @@ def test_extension_array_meta_persists():
     assert table[0].chunks[0].field.metadata_str == input_metadata
 
 
-def test_from_buffer():
-    arr = np.array([1.0, 2.0, 3.0], dtype=np.float64)
-    mv = memoryview(arr)
-    assert pa.array(mv) == pa.array(Array.from_buffer(mv))
-
-    arr = np.array([True, False, True], dtype=np.bool_)
-    mv = memoryview(arr)
-    assert pa.array(mv) == pa.array(Array.from_buffer(mv))
-
-    arr = np.array([1, 2, 3], dtype=np.int64)
-    mv = memoryview(arr)
-    assert pa.array(mv) == pa.array(Array.from_buffer(mv))
-
-    # pyarrow applies some casting; this is weird
-    # According to joris, this may be because pyarrow doesn't implement direct import of
-    # buffer protocol objects, and instead infers from `pa.array(list(memoryview()))`
-    # float32 -> float64
-    # int32 -> int64
-    # uint64 -> int64
-
-    arr = np.array([1.0, 2.0, 3.0], dtype=np.float32)
-    assert pa.array(Array.from_buffer(memoryview(arr))).type == pa.float32()
-
-    arr = np.array([1, 2, 3], dtype=np.int32)
-    assert pa.array(Array.from_buffer(memoryview(arr))).type == pa.int32()
-
-    arr = np.array([1, 2, 3], dtype=np.int64)
-    assert pa.array(Array.from_buffer(memoryview(arr))).type == pa.int64()
-
-    arr = np.array([1, 2, 3], dtype=np.uint64)
-    assert pa.array(Array.from_buffer(memoryview(arr))).type == pa.uint64()
-
-
 def test_getitem():
     arr = Array([1, 2, 3], DataType.int16())
     assert arr[0].as_py() == 1
